services/twirole/requirements.txt
nltk==3.4.5
numpy
pandas
Pillow>=7.1.0
scikit-learn==0.23.2
torch==1.4.0
torchvision==0.2.1
twarc==1.7.4
urllib3==1.24.2
wordninja==2.0.0
termcolor==1.1.0
Cython==0.29.21

